fp64 recommended

+++ example.py
@@ -5,7 +5,7 @@ from mace.calculators import mace_off
 
 atoms = ase.io.read("static/aspirin.xyz", format="xyz")
 
-calculator = mace_off(model="compiled_models/EGRET_1.model", default_dtype="float32")
+calculator = mace_off(model="compiled_models/EGRET_1.model", default_dtype="float64")
 
 calculator.calculate(atoms, ["energy", "forces"], all_changes)
